Move boilerplate code out of integration_tests
